requirements_simple.txt
streamlit
python-dotenv
google-generativeai
chromadb
pypdf
python-docx
openpyxl
sentence-transformers